Make add view work
Add view now works!
With a simpler implementation than before!

The Sighting model was changed
Now enforce unique squirrel id to be unique
Added a Regex validator for the field Unique Squirrel ID. Otherwise the
detail view wouldn't work, because we are matching squirrel id with a
regex.

Unique squirrel ID in the csv file was not actually unqiue, so I
dropped duplicates in the import_squirrel_data command

The 'Date' field is now changed to a DateField